chore: rename emission-type to reference
